refactor(provider): split provider.go into aliases / defaults / factory
provider.go had grown to ~600 lines mixing four concerns: the alias registry, config defaults / merging, the dispatch logic that picks a concrete provider client, and the public Provider interfaces. This commit splits it into four focused files with a short package doc comment that explains the layout, while preserving every symbol and its visibility:

  * provider.go    - public Provider interfaces and the New / NewWithModels entry points only.

  * aliases.go     - Alias struct, the Aliases table, LookupAlias, EachAlias, and the AllProviders / IsKnownProvider / CatalogProviders / IsCatalogProvider helpers built on top of it.

  * defaults.go    - applyProviderDefaults and its helpers (mergeFromProviderConfig, applyAliasFallbacks, setIfNil, setProviderOptIfAbsent, cloneModelConfig), the model-defaults logic (applyModelDefaults, ensureInterleavedThinking), and the model-name predicates (isOpenAIThinkingOnlyModel, isBedrockClaudeModel, gemini3Family, ...). Also hosts resolveProviderType / resolveEffectiveProvider / isOpenAICompatibleProvider since they participate in the same merge pipeline.

  * factory.go     - createDirectProvider, the providerFactory type and providerFactories registry, the per-provider factory functions, the resolveRoutedModel helper and the createRuleBasedRouter wrapper.

Pure code reorganisation: no behaviour change, no API change. go test ./pkg/model/provider/... and golangci-lint both pass.
