+++ pkg/model/provider/provider.go
@@ -1,175 +1,46 @@
+// Package provider builds and dispatches to LLM provider clients.
+//
+// The package is organised across four files:
+//
+//   - provider.go (this file): the public Provider interfaces and the entry
+//     points [New] and [NewWithModels] that callers use to construct a
+//     provider from a model config.
+//   - aliases.go: the built-in provider alias table (OpenAI-compatible
+//     gateways such as ollama, mistral, xai, ...) and the helpers that expose
+//     it to other packages without leaking the underlying map.
+//   - defaults.go: pure config-merging logic that fills in defaults from
+//     custom providers, built-in aliases, and model-specific rules
+//     (thinking budget, interleaved thinking, ...).
+//   - factory.go: dispatch from a resolved provider type to the concrete
+//     client constructor (openai, anthropic, google, dmr, amazon-bedrock,
+//     vertex AI), plus the rule-based router.
 package provider
 
 import (
 	"context"
-	"fmt"
-	"iter"
 	"log/slog"
-	"maps"
-	"slices"
-	"strings"
 
 	"github.com/docker/docker-agent/pkg/chat"
 	"github.com/docker/docker-agent/pkg/config/latest"
 	"github.com/docker/docker-agent/pkg/environment"
-	"github.com/docker/docker-agent/pkg/model/provider/anthropic"
 	"github.com/docker/docker-agent/pkg/model/provider/base"
-	"github.com/docker/docker-agent/pkg/model/provider/bedrock"
-	"github.com/docker/docker-agent/pkg/model/provider/dmr"
-	"github.com/docker/docker-agent/pkg/model/provider/gemini"
-	"github.com/docker/docker-agent/pkg/model/provider/openai"
 	"github.com/docker/docker-agent/pkg/model/provider/options"
-	"github.com/docker/docker-agent/pkg/model/provider/rulebased"
-	"github.com/docker/docker-agent/pkg/model/provider/vertexai"
 	"github.com/docker/docker-agent/pkg/rag/types"
 	"github.com/docker/docker-agent/pkg/tools"
 )
 
-// Alias defines the configuration for a provider alias
-type Alias struct {
-	APIType     string // The actual API type to use (openai, anthropic, etc.)
-	BaseURL     string // Default base URL for the provider
-	TokenEnvVar string // Environment variable name for the API token
-}
-
-// CoreProviders lists all natively implemented provider types.
-// These are the provider types that have direct implementations (not aliases).
-var CoreProviders = []string{
-	"openai",
-	"anthropic",
-	"google",
-	"dmr",
-	"amazon-bedrock",
-}
-
-// AllProviders returns all known provider names (core providers + aliases),
-// sorted for deterministic output.
-func AllProviders() []string {
-	providers := slices.Concat(CoreProviders, slices.Collect(maps.Keys(Aliases)))
-	slices.Sort(providers)
-	return providers
-}
-
-// IsKnownProvider returns true if the provider name is a core provider or an alias.
-func IsKnownProvider(name string) bool {
-	if slices.Contains(CoreProviders, strings.ToLower(name)) {
-		return true
-	}
-	_, exists := LookupAlias(strings.ToLower(name))
-	return exists
-}
-
-// CatalogProviders returns the list of provider names that should be shown in the model catalog.
-// This includes core providers and aliases that have a defined BaseURL (self-contained endpoints).
-// Aliases without a BaseURL (like azure) require user configuration and are excluded.
-func CatalogProviders() []string {
-	providers := make([]string, 0, len(CoreProviders)+len(Aliases))
-
-	// Add all core providers
-	providers = append(providers, CoreProviders...)
-
-	// Add aliases that have a defined BaseURL (they work out of the box)
-	for name, alias := range EachAlias() {
-		if alias.BaseURL != "" {
-			providers = append(providers, name)
-		}
-	}
-
-	return providers
-}
-
-// IsCatalogProvider returns true if the provider name is valid for the model catalog.
-func IsCatalogProvider(name string) bool {
-	// Check core providers
-	if slices.Contains(CoreProviders, name) {
-		return true
-	}
-	// Check aliases with BaseURL
-	if alias, exists := LookupAlias(name); exists && alias.BaseURL != "" {
-		return true
-	}
-	return false
-}
-
-// Aliases maps provider names to their corresponding configurations.
-//
-// Most consumers should call [LookupAlias] for a single lookup or [EachAlias]
-// to iterate, both of which keep the rest of the codebase decoupled from this
-// concrete map. Direct mutation of Aliases is not supported.
-var Aliases = map[string]Alias{
-	"requesty": {
-		APIType:     "openai",
-		BaseURL:     "https://router.requesty.ai/v1",
-		TokenEnvVar: "REQUESTY_API_KEY",
-	},
-	"azure": {
-		APIType:     "openai",
-		TokenEnvVar: "AZURE_API_KEY",
-	},
-	"xai": {
-		APIType:     "openai",
-		BaseURL:     "https://api.x.ai/v1",
-		TokenEnvVar: "XAI_API_KEY",
-	},
-	"nebius": {
-		APIType:     "openai",
-		BaseURL:     "https://api.studio.nebius.com/v1",
-		TokenEnvVar: "NEBIUS_API_KEY",
-	},
-	"mistral": {
-		APIType:     "openai",
-		BaseURL:     "https://api.mistral.ai/v1",
-		TokenEnvVar: "MISTRAL_API_KEY",
-	},
-	"ollama": {
-		APIType: "openai",
-		BaseURL: "http://localhost:11434/v1",
-	},
-	"minimax": {
-		APIType:     "openai",
-		BaseURL:     "https://api.minimax.io/v1",
-		TokenEnvVar: "MINIMAX_API_KEY",
-	},
-	"github-copilot": {
-		APIType:     "openai",
-		BaseURL:     "https://api.githubcopilot.com",
-		TokenEnvVar: "GITHUB_TOKEN",
-	},
-}
-
-// LookupAlias returns the Alias registered for the given name (if any).
-// Lookup is case-sensitive; callers that need case-insensitive matching
-// should normalise the name first (e.g. [strings.ToLower]).
-func LookupAlias(name string) (Alias, bool) {
-	alias, ok := Aliases[name]
-	return alias, ok
-}
-
-// EachAlias returns an iterator over every registered (name, Alias) pair.
-// Iteration order is not guaranteed; callers that need a deterministic order
-// should sort by name.
-func EachAlias() iter.Seq2[string, Alias] {
-	return func(yield func(string, Alias) bool) {
-		for name, alias := range Aliases {
-			if !yield(name, alias) {
-				return
-			}
-		}
-	}
-}
-
-// Provider defines the interface for model providers
+// Provider defines the interface for model providers.
 type Provider interface {
 	// ID returns the model provider ID
 	ID() string
-	// CreateChatCompletionStream creates a streaming chat completion request
-	// It returns a stream that can be iterated over to get completion chunks
+	// CreateChatCompletionStream creates a streaming chat completion request.
+	// It returns a stream that can be iterated over to get completion chunks.
 	CreateChatCompletionStream(
 		ctx context.Context,
 		messages []chat.Message,
 		tools []tools.Tool,
 	) (chat.MessageStream, error)
-	// BaseConfig returns the base configuration of this provider
+	// BaseConfig returns the base configuration of this provider.
 	BaseConfig() base.Config
 }
 
@@ -217,403 +88,4 @@ func NewWithModels(ctx context.Context, cfg *latest.ModelConfig, models map[stri
 	}
 
 	return createDirectProvider(ctx, cfg, env, opts...)
-}
-
-// createRuleBasedRouter creates a rule-based routing provider.
-func createRuleBasedRouter(ctx context.Context, cfg *latest.ModelConfig, models map[string]latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return rulebased.NewClient(ctx, cfg, models, env, resolveRoutedModel, opts...)
-}
-
-// resolveRoutedModel is the rulebased.ProviderFactory used by
-// createRuleBasedRouter. It resolves a routing target — which is either a name
-// from the models map or an inline "provider/model" spec — and returns the
-// provider for it. Routing targets cannot themselves have routing rules.
-//
-// Defined as a package-level function (rather than an inline closure) so the
-// recursion-prevention, parse-error and factory-error paths can be unit-tested
-// directly without going through rulebased.NewClient.
-func resolveRoutedModel(
-	ctx context.Context,
-	modelSpec string,
-	models map[string]latest.ModelConfig,
-	env environment.Provider,
-	factoryOpts ...options.Opt,
-) (rulebased.Provider, error) {
-	// Check if modelSpec is a reference to a model in the models map.
-	if modelCfg, exists := models[modelSpec]; exists {
-		// Prevent infinite recursion - referenced models cannot have routing rules.
-		if len(modelCfg.Routing) > 0 {
-			return nil, fmt.Errorf("model %q has routing rules and cannot be used as a routing target", modelSpec)
-		}
-		return createDirectProvider(ctx, &modelCfg, env, factoryOpts...)
-	}
-
-	// Otherwise, treat as an inline model spec (e.g., "openai/gpt-4o").
-	inlineCfg, parseErr := latest.ParseModelRef(modelSpec)
-	if parseErr != nil {
-		return nil, fmt.Errorf("invalid model spec %q: expected 'provider/model' format or a model reference", modelSpec)
-	}
-	return createDirectProvider(ctx, &inlineCfg, env, factoryOpts...)
-}
-
-// createDirectProvider creates a provider without routing (direct model access).
-func createDirectProvider(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	var globalOptions options.ModelOptions
-	for _, opt := range opts {
-		opt(&globalOptions)
-	}
-
-	// Apply defaults from custom providers (from config) or built-in aliases
-	enhancedCfg := applyProviderDefaults(cfg, globalOptions.Providers())
-
-	providerType := resolveProviderType(enhancedCfg)
-
-	factory, ok := providerFactories[providerType]
-	if !ok {
-		slog.Error("Unknown provider type", "type", providerType)
-		return nil, fmt.Errorf("unknown provider type: %s", providerType)
-	}
-	return factory(ctx, enhancedCfg, env, opts...)
-}
-
-// providerFactory builds a Provider from a fully-resolved ModelConfig.
-// Tests may swap entries in providerFactories to exercise dispatch logic
-// without spinning up real provider clients.
-type providerFactory func(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error)
-
-// providerFactories maps a resolved provider type (the value returned by
-// resolveProviderType) to its constructor. The map is package-private but
-// modifiable; tests must restore the original entries with t.Cleanup.
-var providerFactories = map[string]providerFactory{
-	"openai":                 openaiFactory,
-	"openai_chatcompletions": openaiFactory,
-	"openai_responses":       openaiFactory,
-	"anthropic":              anthropicFactory,
-	"google":                 googleFactory,
-	"dmr":                    dmrFactory,
-	"amazon-bedrock":         bedrockFactory,
-}
-
-func openaiFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return openai.NewClient(ctx, cfg, env, opts...)
-}
-
-func anthropicFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return anthropic.NewClient(ctx, cfg, env, opts...)
-}
-
-func googleFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	// Route non-Gemini models on Vertex AI (Model Garden) through the
-	// vertexai package, which picks the right endpoint per publisher.
-	if vertexai.IsModelGardenConfig(cfg) {
-		return vertexClientFactory(ctx, cfg, env, opts...)
-	}
-	return geminiClientFactory(ctx, cfg, env, opts...)
-}
-
-// geminiClientFactory and vertexClientFactory are the inner constructors used
-// by googleFactory. They are package-level variables (rather than direct
-// references to gemini.NewClient / vertexai.NewClient) so that tests can swap
-// them with fakes via t.Cleanup and assert that googleFactory routes correctly
-// based on vertexai.IsModelGardenConfig — without spinning up real clients.
-var (
-	geminiClientFactory providerFactory = func(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-		return gemini.NewClient(ctx, cfg, env, opts...)
-	}
-	vertexClientFactory providerFactory = func(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-		return vertexai.NewClient(ctx, cfg, env, opts...)
-	}
-)
-
-func dmrFactory(ctx context.Context, cfg *latest.ModelConfig, _ environment.Provider, opts ...options.Opt) (Provider, error) {
-	return dmr.NewClient(ctx, cfg, opts...)
-}
-
-func bedrockFactory(ctx context.Context, cfg *latest.ModelConfig, env environment.Provider, opts ...options.Opt) (Provider, error) {
-	return bedrock.NewClient(ctx, cfg, env, opts...)
-}
-
-// ---------------------------------------------------------------------------
-// Provider-type resolution
-// ---------------------------------------------------------------------------
-
-// resolveProviderType determines the effective API type for a config.
-// Priority: ProviderOpts["api_type"] > built-in alias > provider name.
-func resolveProviderType(cfg *latest.ModelConfig) string {
-	if cfg.ProviderOpts != nil {
-		if apiType, ok := cfg.ProviderOpts["api_type"].(string); ok && apiType != "" {
-			return apiType
-		}
-	}
-	if alias, exists := LookupAlias(cfg.Provider); exists && alias.APIType != "" {
-		return alias.APIType
-	}
-	return cfg.Provider
-}
-
-// ---------------------------------------------------------------------------
-// Provider defaults
-// ---------------------------------------------------------------------------
-
-// applyProviderDefaults applies default configuration from custom providers or built-in aliases.
-// Custom providers (from config) take precedence over built-in aliases.
-// This sets default base URLs, token keys, api_type, and model-specific defaults (like thinking budget).
-//
-// The returned config is a deep-enough copy: the caller's ModelConfig, ProviderOpts map,
-// and ThinkingBudget/TaskBudget pointers are never mutated.
-func applyProviderDefaults(cfg *latest.ModelConfig, customProviders map[string]latest.ProviderConfig) *latest.ModelConfig {
-	// Create a copy to avoid modifying the original.
-	// cloneModelConfig also deep-copies ProviderOpts so writes are safe.
-	enhancedCfg := cloneModelConfig(cfg)
-
-	if providerCfg, exists := customProviders[cfg.Provider]; exists {
-		slog.Debug("Applying custom provider defaults",
-			"provider", cfg.Provider,
-			"model", cfg.Model,
-			"base_url", providerCfg.BaseURL,
-		)
-		mergeFromProviderConfig(enhancedCfg, providerCfg)
-		applyModelDefaults(enhancedCfg)
-		return enhancedCfg
-	}
-
-	if alias, exists := LookupAlias(cfg.Provider); exists {
-		applyAliasFallbacks(enhancedCfg, alias)
-	}
-
-	// Apply model-specific defaults (e.g., thinking budget for Claude/GPT models)
-	applyModelDefaults(enhancedCfg)
-	return enhancedCfg
-}
-
-// mergeFromProviderConfig merges defaults from a custom ProviderConfig into a
-// ModelConfig. Model-level fields always take precedence; provider-level fields
-// only fill in unset (nil/empty) fields. The destination ProviderOpts map is
-// assumed to be safe to mutate (cloneModelConfig copies it).
-func mergeFromProviderConfig(dst *latest.ModelConfig, src latest.ProviderConfig) {
-	// Apply the underlying provider type if set on the provider config.
-	// This allows the model to inherit the real provider type (e.g., "anthropic")
-	// so that the correct API client is selected.
-	if src.Provider != "" {
-		dst.Provider = src.Provider
-	}
-
-	if dst.BaseURL == "" {
-		dst.BaseURL = src.BaseURL
-	}
-	if dst.TokenKey == "" {
-		dst.TokenKey = src.TokenKey
-	}
-	setIfNil(&dst.Temperature, src.Temperature)
-	setIfNil(&dst.MaxTokens, src.MaxTokens)
-	setIfNil(&dst.TopP, src.TopP)
-	setIfNil(&dst.FrequencyPenalty, src.FrequencyPenalty)
-	setIfNil(&dst.PresencePenalty, src.PresencePenalty)
-	setIfNil(&dst.ParallelToolCalls, src.ParallelToolCalls)
-	setIfNil(&dst.TrackUsage, src.TrackUsage)
-	setIfNil(&dst.ThinkingBudget, src.ThinkingBudget)
-	setIfNil(&dst.TaskBudget, src.TaskBudget)
-
-	// Merge provider_opts from provider config (model opts take precedence).
-	for k, v := range src.ProviderOpts {
-		if dst.ProviderOpts == nil {
-			dst.ProviderOpts = make(map[string]any)
-		}
-		if _, has := dst.ProviderOpts[k]; !has {
-			dst.ProviderOpts[k] = v
-		}
-	}
-
-	// Set api_type in ProviderOpts if not already set.
-	// Prefer the explicit APIType from the provider config; otherwise default to
-	// openai_chatcompletions for OpenAI-compatible providers.
-	switch {
-	case src.APIType != "":
-		setProviderOptIfAbsent(dst, "api_type", src.APIType)
-	case isOpenAICompatibleProvider(resolveEffectiveProvider(src)):
-		setProviderOptIfAbsent(dst, "api_type", "openai_chatcompletions")
-	}
-}
-
-// applyAliasFallbacks applies BaseURL and TokenKey defaults from a built-in
-// alias when the model config does not already specify them.
-func applyAliasFallbacks(dst *latest.ModelConfig, alias Alias) {
-	if dst.BaseURL == "" {
-		dst.BaseURL = alias.BaseURL
-	}
-	if dst.TokenKey == "" {
-		dst.TokenKey = alias.TokenEnvVar
-	}
-}
-
-// setIfNil assigns src to *dst when *dst is nil. It centralises the repetitive
-// "only fill in if unset" pattern used when merging provider-level defaults.
-func setIfNil[T any](dst **T, src *T) {
-	if *dst == nil && src != nil {
-		*dst = src
-	}
-}
-
-// setProviderOptIfAbsent stores key=value in cfg.ProviderOpts unless the key is
-// already set. The map is created lazily.
-func setProviderOptIfAbsent(cfg *latest.ModelConfig, key string, value any) {
-	if cfg.ProviderOpts == nil {
-		cfg.ProviderOpts = make(map[string]any)
-	}
-	if _, has := cfg.ProviderOpts[key]; !has {
-		cfg.ProviderOpts[key] = value
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Thinking defaults and overrides
-// ---------------------------------------------------------------------------
-
-// applyModelDefaults applies provider-specific default values for model configuration.
-//
-// Thinking defaults policy:
-//   - thinking_budget: 0  or  thinking_budget: none  →  thinking is off (nil).
-//   - thinking_budget explicitly set to a real value  →  kept as-is; interleaved_thinking
-//     is auto-enabled for Anthropic/Bedrock-Claude.
-//   - thinking_budget NOT set:
-//   - Thinking-only models (OpenAI o-series) get "medium".
-//   - All other models get no thinking.
-//
-// NOTE: max_tokens is NOT set here; see teamloader and runtime/model_switcher.
-func applyModelDefaults(cfg *latest.ModelConfig) {
-	// Explicitly disabled → normalise to nil so providers never see it.
-	if cfg.ThinkingBudget.IsDisabled() {
-		cfg.ThinkingBudget = nil
-		slog.Debug("Thinking explicitly disabled",
-			"provider", cfg.Provider, "model", cfg.Model)
-		return
-	}
-
-	providerType := resolveProviderType(cfg)
-
-	// User already set a real thinking_budget — just apply side-effects.
-	if cfg.ThinkingBudget != nil {
-		ensureInterleavedThinking(cfg, providerType)
-		return
-	}
-
-	// No thinking_budget configured — only thinking-only models get a default.
-	switch providerType {
-	case "openai", "openai_chatcompletions", "openai_responses":
-		if isOpenAIThinkingOnlyModel(cfg.Model) {
-			cfg.ThinkingBudget = &latest.ThinkingBudget{Effort: "medium"}
-			slog.Debug("Applied default thinking for thinking-only OpenAI model",
-				"provider", cfg.Provider, "model", cfg.Model)
-		}
-	}
-}
-
-// ---------------------------------------------------------------------------
-// Shared helpers
-// ---------------------------------------------------------------------------
-
-// cloneModelConfig returns a shallow copy of cfg with a deep copy of
-// ProviderOpts so that callers can safely mutate the returned config's
-// map and pointer fields without affecting the original.
-func cloneModelConfig(cfg *latest.ModelConfig) *latest.ModelConfig {
-	c := *cfg
-	if cfg.ProviderOpts != nil {
-		c.ProviderOpts = make(map[string]any, len(cfg.ProviderOpts))
-		maps.Copy(c.ProviderOpts, cfg.ProviderOpts)
-	}
-	return &c
-}
-
-// ensureInterleavedThinking sets interleaved_thinking=true in ProviderOpts
-// for Anthropic and Bedrock-Claude models, unless the user already set it.
-func ensureInterleavedThinking(cfg *latest.ModelConfig, providerType string) {
-	needsInterleaved := providerType == "anthropic" ||
-		(providerType == "amazon-bedrock" && isBedrockClaudeModel(cfg.Model))
-	if !needsInterleaved {
-		return
-	}
-	if cfg.ProviderOpts == nil {
-		cfg.ProviderOpts = make(map[string]any)
-	}
-	if _, has := cfg.ProviderOpts["interleaved_thinking"]; !has {
-		cfg.ProviderOpts["interleaved_thinking"] = true
-		slog.Debug("Auto-enabled interleaved_thinking",
-			"provider", cfg.Provider, "model", cfg.Model)
-	}
-}
-
-// isOpenAIThinkingOnlyModel returns true for OpenAI models that require thinking
-// to function properly (o-series reasoning models).
-func isOpenAIThinkingOnlyModel(model string) bool {
-	m := strings.ToLower(model)
-	return strings.HasPrefix(m, "o1") ||
-		strings.HasPrefix(m, "o3") ||
-		strings.HasPrefix(m, "o4")
-}
-
-// isBedrockClaudeModel returns true if the model ID is a Claude model on Bedrock.
-// Claude model IDs on Bedrock start with "anthropic.claude-" or "global.anthropic.claude-".
-func isBedrockClaudeModel(model string) bool {
-	m := strings.ToLower(model)
-	return strings.HasPrefix(m, "anthropic.claude-") || strings.HasPrefix(m, "global.anthropic.claude-")
-}
-
-// gemini3Family extracts the model family (e.g. "pro", "flash") from a
-// Gemini 3+ model name, or returns "" if the model is not Gemini 3+.
-// It handles both "gemini-3-<family>" and "gemini-3.X-<family>" patterns.
-//
-// Examples:
-//
-//	gemini3Family("gemini-3-pro")              → "pro"
-//	gemini3Family("gemini-3.1-flash-preview")  → "flash-preview"
-//	gemini3Family("gemini-2.5-flash")          → ""
-func gemini3Family(model string) string {
-	if !strings.HasPrefix(model, "gemini-3") {
-		return ""
-	}
-	rest := model[len("gemini-3"):]
-	if rest == "" {
-		return ""
-	}
-	// Accept "gemini-3-..." or "gemini-3.X-..." (e.g. gemini-3.1-pro-preview)
-	switch rest[0] {
-	case '-':
-		return rest[1:] // "gemini-3-pro" → "pro"
-	case '.':
-		if _, family, ok := strings.Cut(rest, "-"); ok {
-			return family // "gemini-3.1-pro-preview" → "pro-preview"
-		}
-	}
-	return ""
-}
-
-func isGeminiProModel(model string) bool {
-	return strings.HasPrefix(gemini3Family(model), "pro")
-}
-
-func isGeminiFlashModel(model string) bool {
-	return strings.HasPrefix(gemini3Family(model), "flash")
-}
-
-// resolveEffectiveProvider returns the effective provider type for a ProviderConfig.
-// If Provider is explicitly set, returns that. Otherwise returns "openai" (backward compat).
-func resolveEffectiveProvider(cfg latest.ProviderConfig) string {
-	if cfg.Provider != "" {
-		return cfg.Provider
-	}
-	return "openai"
-}
-
-// isOpenAICompatibleProvider returns true if the provider type uses the OpenAI API protocol.
-func isOpenAICompatibleProvider(providerType string) bool {
-	switch providerType {
-	case "openai", "openai_chatcompletions", "openai_responses":
-		return true
-	default:
-		// Check if it's an alias that maps to openai
-		if alias, exists := LookupAlias(providerType); exists {
-			return alias.APIType == "openai"
-		}
-		return false
-	}
 }
